backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Sistema de reservas de salas com notificações assíncronas por e-mail utilizando o padrão Outbox"
authors = [
    {name = "mateusb12",email = "matbessa12@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "pydantic (>=2.13.1,<3.0.0)",
    "repomix (>=0.5.0,<0.6.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.scripts]
clean-ai-comments = "clean_ai_comments.cleaner:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "pre-commit (>=4.3.0,<5.0.0)",
    "libcst (>=1.8.6,<2.0.0)",
    "black (>=24.4.2,<25.0.0)",
    "ruff (>=0.14.5,<0.15.0)"
]

[tool.ruff]
line-length = 140
exclude = [
    "alembic",
    ".venv",
    ".ruff_cache",
    "__pycache__",
]

[tool.ruff.lint]
preview = true
extend-select = [
    "E",
    "F",
    "W",
    "C90",
]
ignore = [
    "E501",
    "E302",
    "W292",
    "D203",
    "D212",
    "D213",
    "D100",
    "D101",
    "D102",
    "D205",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
docstring-code-format = true
docstring-code-line-length = 140

[tool.ruff.lint.mccabe]
max-complexity = 20

[tool.black]
line-length = 140
target-version = ["py312"]
skip-string-normalization = true
exclude = '''
(
  alembic
  | \.venv
  | \.ruff_cache
  | __pycache__
)
'''

[tool.pytest.ini_options]
pythonpath = ["."]
